Entity Framework eklendi Brand ve Color nesneleri eklendi Sql server tarafında veritabanı kuruldu, tablolar oluşturuldu Sisteme Generic IEntitiyRepository class'ı eklendi Nesneler Entity Framework altyapısıyla yazıldı BrandId ve ColorId'ye göre araba getirme işi yapıldı Veritabanına yeni araba eklenirken - Arabanın adı en az 2 kelime - Günlük ödeme 0'dan yukarı Olacak şekilde ayarlandı

+++ Entities/Concrete/Car.cs
@@ -10,8 +10,8 @@ namespace Entities.Concrete
         public int Id { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public short ModelYear { get; set; }
-        public float DailyPrice { get; set; }
+        public int ModelYear { get; set; }
+        public double DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
